+++ next-app/src/components/views/Introduce/IntroduceHeaderTitle.tsx
@@ -1,0 +1,12 @@
+import * as S from 'components/views/Introduce/Introduce.style'
+
+const IntroduceHeaderTitle = () => {
+  return (
+    <S.HeaderTitle>
+      여기저기 둥둥💭 떠있는 나의 인사이트💡 콘텐츠📚를{' '}
+      <S.PinkFeedoong>피둥🐽</S.PinkFeedoong>으로 모아보세요.
+    </S.HeaderTitle>
+  )
+}
+
+export default IntroduceHeaderTitle
