--- next-app/src/components/views/Introduce/Top/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './Top'
